splitting windows tests
